cleanup nn_lib.py, fix all models for multiboard, mbv2
